requirements.txt
Django==3.1.5
requests==2.25.1
behave==1.2.6
djangorestframework==3.12.2
pinject==0.14.1
psutil==5.8.0
pytest==6.2.2
